refactor: relocate document type
