+++ runtime/runtime-core/common/src/aws/smithy/kotlin/runtime/content/DocumentBuilder.kt
@@ -2,7 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package aws.smithy.kotlin.runtime.smithy
+package aws.smithy.kotlin.runtime.content
 
 import kotlin.jvm.JvmName
